Implemented most of admin control panel, missing post controls.

+++ main/user/deleteUser.php
@@ -1,17 +1,25 @@
 <?php
 require_once('../../_libs/json.php');
 require_once('../../_libs/csv.php');
+require_once('../../_libs/auth.php');
 require_once('../classes/DBInterface.php');
 require_once('../classes/User.php');
 require_once('../settings.php');
 
+session_start();
+
+if(!(is_user($_SESSION['user_id'], $_GET['id']))){
+    header('Location: ../detail.php?id='.$_GET['id']);
+    die();
+}
 
 deleteUser($_GET['id']);
 
 function deleteUser($u_id){
     $db = DBInterface::connectToDB(DB_SETTINGS, DB_OPTIONS);
-    $delete_query = $db->prepare("DELETE FROM user WHERE user_id=?");
+    $delete_query = $db->prepare("UPDATE user SET role=0 WHERE user_id=?");
     $delete_query->execute([$u_id]);
+    $_SESSION = [];
     header('Location: ../index.php');
     die();
 }
